+++ GNUmakefile
@@ -425,3 +425,8 @@ endif
 missing: ## Check for packages not being tested
 	@echo "==> Checking for packages not being tested ..."
 	@go run -modfile tools/go.mod tools/missing/main.go .github/workflows/test-core.yaml
+
+.PHONY: ec2info
+ec2info: ## Generate AWS EC2 CPU specification table
+	@echo "==> Generating AWS EC2 specifications ..."
+	@go run -modfile tools/go.mod tools/ec2info/main.go
